requirements.txt
boto3>=1.28.0
streamlit>=1.24.0
python-dotenv>=1.0.0 